# sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NexusRAG API",
    "version": "v1"
  },
  "paths": {
...
    "/v1/audit/events": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "List Audit Events",
        "operationId": "list_audit_events_v1_audit_events_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenant Id"
            }
          },
          {
            "name": "event_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Type"
            }
          },
          {
            "name": "outcome",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Outcome"
            }
          },
          {
            "name": "resource_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resource Type"
            }
          },
          {
            "name": "resource_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resource Id"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_AuditEventsPage_"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "BAD_REQUEST",
                    "message": "Bad request"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "AUTH_UNAUTHORIZED",
                    "message": "Missing or invalid bearer token"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "402": {
            "description": "Quota exceeded",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "QUOTA_EXCEEDED",
                    "message": "Monthly request quota exceeded",
                    "details": {
                      "period": "month",
                      "limit": 10000,
                      "used": 10000,
                      "remaining": 0
                    }
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "AUTH_FORBIDDEN",
                    "message": "Insufficient role for this operation"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "NOT_FOUND",
                    "message": "Resource not found"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "IDEMPOTENCY_KEY_CONFLICT",
                    "message": "Idempotency-Key already used with different payload"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "REQUEST_VALIDATION_ERROR",
                    "message": "Validation error"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "RATE_LIMITED",
                    "message": "Rate limit exceeded",
                    "details": {
                      "scope": "api_key",
                      "route_class": "read",
                      "retry_after_ms": 1200
                    }
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INTERNAL_ERROR",
                    "message": "Internal server error"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SERVICE_UNAVAILABLE",
                    "message": "Service unavailable"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/audit/events/{event_id}": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Get Audit Event",
        "operationId": "get_audit_event_v1_audit_events__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_AuditEventResponse_"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "BAD_REQUEST",
                    "message": "Bad request"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "AUTH_UNAUTHORIZED",
                    "message": "Missing or invalid bearer token"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "402": {
            "description": "Quota exceeded",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "QUOTA_EXCEEDED",
                    "message": "Monthly request quota exceeded",
                    "details": {
                      "period": "month",
                      "limit": 10000,
                      "used": 10000,
                      "remaining": 0
                    }
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "AUTH_FORBIDDEN",
                    "message": "Insufficient role for this operation"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "NOT_FOUND",
                    "message": "Resource not found"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "IDEMPOTENCY_KEY_CONFLICT",
                    "message": "Idempotency-Key already used with different payload"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "REQUEST_VALIDATION_ERROR",
                    "message": "Validation error"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "RATE_LIMITED",
                    "message": "Rate limit exceeded",
                    "details": {
                      "scope": "api_key",
                      "route_class": "read",
                      "retry_after_ms": 1200
                    }
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INTERNAL_ERROR",
                    "message": "Internal server error"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SERVICE_UNAVAILABLE",
                    "message": "Service unavailable"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditEventResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "occurred_at": {
            "type": "string",
            "title": "Occurred At"
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          },
          "actor_type": {
            "type": "string",
            "title": "Actor Type"
          },
          "actor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Id"
          },
          "actor_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Role"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "outcome": {
            "type": "string",
            "title": "Outcome"
          },
          "resource_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Type"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "ip_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Address"
          },
          "user_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent"
          },
          "metadata_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Json"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "occurred_at",
          "tenant_id",
          "actor_type",
          "actor_id",
          "actor_role",
          "event_type",
          "outcome",
          "resource_type",
          "resource_id",
          "request_id",
          "ip_address",
          "user_agent",
          "metadata_json",
          "error_code",
          "created_at"
        ],
        "title": "AuditEventResponse"
      },
      "AuditEventsPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditEventResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "next_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_offset"
        ],
        "title": "AuditEventsPage"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "error",
          "meta"
        ],
        "title": "ErrorEnvelope"
      },
...
      "ResponseMeta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "api_version": {
            "type": "string",
            "title": "Api Version",
            "default": "v1"
          }
        },
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "ResponseMeta"
      },
...
      "SuccessEnvelope_AuditEventResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AuditEventResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessEnvelope[AuditEventResponse]"
      },
      "SuccessEnvelope_AuditEventsPage_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AuditEventsPage"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessEnvelope[AuditEventsPage]"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}